Some CSS tweaks to make slideup/slidedown/pop/fade transitions run smoother on the Motorola Xoom (Android 3.1). Basically, we needed to specify translate3d() on the page that does *NOT* animate to avoid flashing. This basically pushes the non-animating page into the same rendering pipeline as the one being animated and avoids a pre-flash.
Note that things *STILL* blink badly on the Xoom after the transition completes. This is due to the removal of the ui-page-active class on the "from" page which toggles its display to display:none. Apparently the Xoom redraws all positioned elements when the display of one of them is turned off. Still exploring ways around this including off-screen placement which seems to be effective, but may have accessibility impact.